Enemy spawner
Enemy spawner can spawn a variable number of any actor class, when a variable number of that enemy class remains in the map. An EnemyCounter must be present and set up correctly in order for the spawner to work. Instructions on how to use these 2 objects can be found as comments in their blueprints