Attack system linked with animations. Attacks do damages
The attack system is now licked to the animation blueprint of the character allowing animations to be player from player inputs. All animations have been set up to do damage, prevent movement, prevent dodging and to reset to the default index number for walking automatically if the combo is dropped. Some values need tweaking to make gameplay feel smoother and code needs cleaning for the final commit