Enemy aggression bug fix
Fixed the bug where the number of aggressive enemies was inaccurate. Dead enemies were still being counted, and so was the boss. The counter is temporarily set to only work with melee enemies, while the boss is being reworked. KNOWN ISSUE: the number of aggressive enemies can be higher than the maximum if enemies are destroyed while more are spawning