Moving Car 1
Added a car that moves along a spline path with the use of 2 new BP's: BP_MovingCar and BP_CarRoute. BP_CarRoute is for setting the spline path that the car will follow. BP_MovingCar is the car itself, which contains all the logic for moving the car along the path. Timings and route are fully customisable.