New Bridge Model + BP_MovingCar Adjustments + New Moving Car & Material
Replaced old instances of bridges with new one from 3D modeller that is scaled more appropriately. Adapted BP_MovingCar to vary TimeBeforeReplay within a small range as well as added comments to this BP. Added a 3rd moving car. Created a new blue material to apply to the pillar closest to the intersection.