Megaera Playtesting Fixes
 The 'Shockwave' produced during the slam attack is moved closer to the ground.
The 'Charged Laser' warning time is adjusted to be easier to dodge.
The 2nd phase 'Spinning Lasers' are made slower so that the player can still outrun them.
The boss moves higher off the ground before the slam attack.
The attack order is changed to as follows: Spinning Lasers > Charged Laser > Waves > Shockwave. The bullet-hell attacks should also swap: BH2 > BH1.
The boss should enter the idle state for ~3 seconds after returning to the ground following the bullet-hell attacks.