Outline Bar New Shape + Colour Change + Incremental Healing
Outline bar is now rectangle, using imported image rather than created shape in material. The colour changes in respect to health percentage. Healing looked buggy, incremental healing means the bar will only increase after the player has stopped killing enemies for a certain amount of time. A secondary bar with the prior method may be helpful as the bar will now not update for a long time if the player is on a long streak of defeating enemies. The buggy look to the increase still happens if the player kills multiple enemies in between checks (every 1s) - may need to lower the time between checks.